Add SEO optimization features: sitemap, robots, JSON-LD, and OpenGraph image
