+++ app/robots.ts
@@ -1,0 +1,15 @@
+import { MetadataRoute } from 'next';
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: [
+      {
+        userAgent: '*',
+        allow: '/',
+        disallow: ['/api/', '/_next/'],
+      },
+    ],
+    sitemap:
+      'https://design-resources-for-developers-tau.vercel.app/sitemap.xml',
+  };
+}
