+++ components/json-ld.tsx
@@ -1,0 +1,63 @@
+'use client';
+
+import Script from 'next/script';
+
+type JsonLdProps = {
+  data: Record<string, unknown>;
+};
+
+export function JsonLd({ data }: JsonLdProps) {
+  return (
+    <Script
+      id='json-ld'
+      type='application/ld+json'
+      dangerouslySetInnerHTML={{ __html: JSON.stringify(data) }}
+    />
+  );
+}
+
+type BreadcrumbItem = {
+  name: string;
+  url: string;
+};
+
+type BreadcrumbJsonLdProps = {
+  items: BreadcrumbItem[];
+};
+
+export function BreadcrumbJsonLd({ items }: BreadcrumbJsonLdProps) {
+  const data = {
+    '@context': 'https://schema.org',
+    '@type': 'BreadcrumbList',
+    itemListElement: items.map((item, index) => ({
+      '@type': 'ListItem',
+      position: index + 1,
+      name: item.name,
+      item: item.url,
+    })),
+  };
+
+  return <JsonLd data={data} />;
+}
+
+type ItemListElement = {
+  position: number;
+  name: string;
+  url: string;
+};
+
+type ItemListJsonLdProps = {
+  name: string;
+  items: ItemListElement[];
+};
+
+export function ItemListJsonLd({ name, items }: ItemListJsonLdProps) {
+  const data = {
+    '@context': 'https://schema.org',
+    '@type': 'ItemList',
+    name,
+    itemListElement: items,
+  };
+
+  return <JsonLd data={data} />;
+}
